Switch to use @OverrideRiakClusterConfig and get rid of SingleNodeRiakSparkTest
