+++ connector/src/test/scala/com/basho/riak/spark/rdd/RDDTest.scala
@@ -24,7 +24,7 @@ import org.junit.{Assert, Test}
 case class TSData(latitude: Float, longitude: Float, timestamp: String, user_id: String, gauge1: Int, gauge2: String)
 
 @Category(Array(classOf[RiakCommonTests]))
-class RDDTest extends SingleNodeRiakSparkTest with AbstractRDDTest {
+class RDDTest extends AbstractRDDTest {
   private final val CREATION_INDEX = "creationNo"
 
   protected override val jsonData = Some(
