+++ connector/src/test/scala/com/basho/riak/spark/rdd/StoreRDDTest.scala
@@ -48,7 +48,7 @@ case class ORMDomainObject(
     login: String)
 
 @Category(Array(classOf[RiakCommonTests]))
-class RDDStoreTest extends SingleNodeRiakSparkTest with AbstractRDDTest {
+class RDDStoreTest extends AbstractRDDTest {
 
   val _expectedException: ExpectedException = ExpectedException.none()
